pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyrad-nifti-workflow"
version = "0.2.0"
description = "Clinical radiomics platform backend built on FastAPI with a local PyRadiomics pipeline."
readme = "README.md"
requires-python = ">=3.9,<3.12"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "numpy>=1.24,<2",
  "pandas>=2.0,<3",
  "scipy>=1.10,<2",
  "SimpleITK>=2.3,<3",
  "scikit-learn>=1.3,<2",
  "matplotlib>=3.7,<4",
  "seaborn>=0.13,<1",
  "pyyaml>=6,<7",
  "openpyxl>=3.1,<4",
  "fastapi>=0.115,<1",
  "uvicorn>=0.30,<1",
]

[project.optional-dependencies]
pip-radiomics = [
  "pyradiomics>=3.1.0,<4",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["pyrad_workflow*"]

[tool.setuptools.package-data]
pyrad_workflow = ["static/*"]
